back_python/requirement.txt
APScheduler==3.8.0
click==8.0.1
colorama==0.4.4
Flask==2.0.1
Flask-APScheduler==1.12.2
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
six==1.16.0
SQLAlchemy==1.4.25
tzlocal==2.1
Werkzeug==2.0.1
